python_version/requirements.txt
pymongo
flask_pymongo
gitpython
requests
lizard
pyyaml
pycrypto